playcontrol in carsession

+++ common/src/main/cpp/include/AudioSource.h
@@ -7,7 +7,7 @@ namespace equalizer {
         virtual ~AudioSource() = default;
 
         // Return 1 sample of audio to be played back
-        virtual int16_t getSample() = 0;
+        virtual float getSample() = 0;
 
         // A callback invoked when the audio stream is stopped
         virtual void onPlaybackStopped() = 0;
